Added raptor_new_term_from_counted_string to write a term as N-Triples
(raptor_new_term_from_counted_string): Added

(raptor_term_to_turtle_counted_string, raptor_term_to_turtle_string):
Added references to new function and N-Triples versions below.

(raptor_term_to_counted_string, raptor_term_to_string): Added
reference to new function and Turtle versions above.

+++ src/raptor_internal.h
@@ -692,7 +692,7 @@ RAPTOR_INTERNAL_API const char* raptor_basename(const char *name);
 int raptor_term_print_as_ntriples(const raptor_term *term, FILE* stream);
 
 /* raptor_ntriples.c */
-int raptor_ntriples_parse_term(raptor_world* world, raptor_locator* locator, unsigned char *string, size_t *len_p, raptor_term** term_p);
+size_t raptor_ntriples_parse_term(raptor_world* world, raptor_locator* locator, unsigned char *string, size_t *len_p, raptor_term** term_p);
 
 /* raptor_parse.c */
 raptor_parser_factory* raptor_world_get_parser_factory(raptor_world* world, const char *name);
